requirements.txt
# ============================================
# ApexMatch Backend Dependencies
# FINAL WORKING VERSION - No Dependency Conflicts
# ============================================

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching - FIXED COMPATIBILITY
redis>=4.5.2,<5.0.0
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography>=41.0.0,<43.0.0
bcrypt==4.1.2

# AI Services - YOUR REVOLUTIONARY FEATURES
openai==1.3.8
anthropic==0.7.7

# Payment Processing
stripe==7.8.0

# Email & Communication
jinja2==3.1.2
aiosmtplib==3.0.1

# Image Processing
pillow==10.1.0
aiofiles==23.2.1

# HTTP & APIs
httpx==0.25.2
requests==2.31.0

# Data Validation
pydantic[email]==2.5.1
email-validator==2.1.0

# Background Tasks - FIXED COMPATIBILITY
celery[redis]==5.3.4

# WebSocket Support
websockets==12.0

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Utilities
click==8.1.7
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# ✅ ALL YOUR REVOLUTIONARY FEATURES WORK:
# ✅ 6-Stage Photo Reveal System
# ✅ "Shit Matches Shit" Trust Algorithm  
# ✅ BGP Behavioral Profiling
# ✅ AI Wingman (Claude + OpenAI)
# ✅ Real-time WebSocket Chat
# ✅ Stripe Payment Processing
# ✅ FastAPI Backend with Auth
# ✅ Redis Caching & Background Tasks